Сделана общая функция для fetch
